+++ udp-project/server/server-udp.c
@@ -20,7 +20,7 @@ R                        **relations;		     // Ptr to the list of relations betw
 int                        timeout; //Timeout of the go back n protocol
 int                        server_port;  //Num port where main thread wait connections
 int                        window;    //the dimension of the window(packets in fly)
-double                     prob_loss;
+int                        prob_loss;
 
 
 void thread_death(){
